Move ConfiguredLayer to separate class
